Move wrangler.toml related functions into their own file
