+++ packages/create-cloudflare/src/cli.ts
@@ -13,6 +13,7 @@ import {
 	rectifyPmMismatch,
 } from "helpers/packageManagers";
 import { installWrangler, npmInstall } from "helpers/packages";
+import { updateWranglerToml } from "helpers/wrangler/config";
 import { version } from "../package.json";
 import {
 	gitCommit,
@@ -32,7 +33,7 @@ import {
 	updatePackageName,
 	updatePackageScripts,
 } from "./templates";
-import { installWorkersTypes, updateWranglerToml } from "./workers";
+import { installWorkersTypes } from "./workers";
 import type { C3Args, C3Context } from "types";
 
 const { npm } = detectPackageManager();
